+++ src/components/Table/TableNew.vue
@@ -3,6 +3,8 @@
     ref="scrollWrapperRef"
     class="table"
     @unmount-scroll="unmountScroll"
+    :is-show-vertical-scroll="props.isShowVerticalScroll"
+    :is-show-horizontal-scroll="props.isShowHorizontalScroll"
     :data-testid="`${props.dataTestid}-ScrollWrapper`"
   >
     <table
@@ -22,8 +24,9 @@
           <slot name="head"></slot>
 
           <HeadTableRowNew
-            class="table__search-tr"
             v-if="$slots['search']"
+            class="table__search-tr"
+            ref="searchRowRef"
             :data-testid="`${props.dataTestid}-Search-Row`"
           >
             <TableTh
@@ -56,7 +59,7 @@ import ScrollWrapperNew from '@/components/ScrollWrapper/ScrollWrapperNew.vue';
 import HeadTableRowNew from '@/components/Table/HeadTableRowNew.vue';
 import TableTh from '@/components/Table/TableTh.vue';
 import changeStyleProperties from '@/helpers/change-style-properties';
-import { computed, onMounted, ref } from 'vue';
+import { computed, onMounted, reactive, ref } from 'vue';
 import type {
   ITableEmit,
   ITableProps
@@ -67,14 +70,23 @@ defineOptions({
 });
 
 const props = withDefaults(defineProps<ITableProps>(), {
-  dataTestid: 'Table'
+  dataTestid: 'Table',
+  isShowHorizontalScroll: false,
+  isShowVerticalScroll: false
 });
 
 const emit = defineEmits<ITableEmit>();
 
+const state = reactive<{
+  headHeight: number;
+}>({
+  headHeight: 0
+});
+
 const tableRef = ref<HTMLElement | null>(null);
 const theadRef = ref<HTMLElement | null>(null);
 const tbodyRef = ref<HTMLElement | null>(null);
+const searchRowRef = ref<InstanceType<typeof HeadTableRowNew> | null>(null);
 const scrollWrapperRef = ref<InstanceType<typeof ScrollWrapperNew> | null>(
   null
 );
@@ -105,6 +117,8 @@ const setHeadHeight = () => {
 
   const headHeight = head.getBoundingClientRect().height;
 
+  state.headHeight = headHeight;
+
   changeStyleProperties(
     {
       '--scroll-track-margin-top': `${headHeight}px`
@@ -113,9 +127,32 @@ const setHeadHeight = () => {
   );
 };
 
+/**
+ * Скроллит вначало
+ */
 const scrollToTop = () => {
   if (scrollWrapperRef.value) {
     scrollWrapperRef.value.scrollToTop();
+  }
+};
+
+/**
+ * устанавливает минимальное значение для таблицы, если в ней есть поиск
+ * и не задано минимальное значение
+ */
+const setSearchMinHeight = () => {
+  if (scrollWrapperRef.value && searchRowRef.value) {
+    const style = getComputedStyle(scrollWrapperRef.value.$el);
+    const minHeight = Number(style.minHeight.replace(/\D/g, ''));
+
+    if (minHeight && minHeight > 0) {
+      return;
+    }
+
+    changeStyleProperties(
+      { 'min-height': `${state.headHeight + 108}px` },
+      scrollWrapperRef.value.$el
+    );
   }
 };
 
@@ -138,6 +175,8 @@ onMounted(() => {
   if (tableRef.value) {
     resizeObserver.observe(tableRef.value);
   }
+
+  setSearchMinHeight();
 });
 </script>
 
@@ -145,6 +184,8 @@ onMounted(() => {
 .table {
   --td-vertical-padding: 11.5px;
   --td-horizontal-padding: 8px;
+  --scroll-slot-background-color: var(--table-background-color, var(--white));
+
   &__table {
     position: relative;
     width: 100%;